please read ChangeLog entries of 2006-03-04 and 2006-03-01

+++ ggc/ggc/datamodels/HbA1cValues.java
@@ -115,7 +115,7 @@ public class HbA1cValues
     public float getHbA1c_Method2()
     {
         if (dayCount > 0)
-            return (float)(getAvgBG() / 30 + 2);
+            return (getAvgBG() / 30 + 2);
         else
             return 0;
     }
